--- src/main/java/ku/user/domain/Ranking/domain/RhythmRanking.java
@@ -1,4 +1,0 @@
-package ku.user.domain.Ranking.domain;
-
-public class RhythmRanking {
-}
