+++ src/main/java/ku/user/domain/ranking/controller/SteppingStonesScoreController.java
@@ -1,23 +1,19 @@
-package ku.user.domain.Ranking.controller;
+package ku.user.domain.ranking.controller;
 
-import ku.user.domain.Ranking.domain.SteppingStonesScore;
-import ku.user.domain.Ranking.dto.request.SaveSteppingRequest;
-import ku.user.domain.Ranking.dto.response.GetRankingResponse;
-import ku.user.domain.Ranking.dto.response.SaveSteppingResponse;
-import ku.user.domain.Ranking.service.RankingService;
-import ku.user.domain.Ranking.service.SteppingStonesScoreService;
+import ku.user.domain.ranking.domain.SteppingStonesScore;
+import ku.user.domain.ranking.dto.request.SaveSteppingRequest;
+import ku.user.domain.ranking.dto.response.GetRankingResponse;
+import ku.user.domain.ranking.dto.response.SaveSteppingResponse;
+import ku.user.domain.ranking.service.RankingService;
+import ku.user.domain.ranking.service.SteppingStonesScoreService;
 import ku.user.global.response.ApiResponse;
 import lombok.RequiredArgsConstructor;
-import org.springframework.data.redis.core.RedisTemplate;
-import org.springframework.data.redis.core.ZSetOperations;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
-import java.util.ArrayList;
 import java.util.List;
-import java.util.Set;
 
 @RestController
 @RequiredArgsConstructor
