+++ src/main/java/ku/user/domain/ranking/domain/RhythmRanking.java
@@ -1,0 +1,4 @@
+package ku.user.domain.ranking.domain;
+
+public class RhythmRanking {
+}
