+++ src/main/java/ku/user/domain/ranking/domain/RhythmScore.java
@@ -1,18 +1,18 @@
-package ku.user.domain.Ranking.domain;
+package ku.user.domain.ranking.domain;
 
 import jakarta.persistence.*;
 import lombok.*;
 import org.hibernate.annotations.CreationTimestamp;
 
 import java.time.LocalDateTime;
-
+// 점수, 캐릭터명, 날짜, 재화
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
 @Builder
 @Getter
-@Table(name = "stepping_stones_score")
-public class SteppingStonesScore {
+@Table(name = "rhythm_scores")
+public class RhythmScore {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
@@ -23,9 +23,6 @@ public class SteppingStonesScore {
     @Column(nullable = false)
     private int score;
 
-    @Column(nullable = false)
-    private int coin;
-
     @Column(name = "created_at", nullable = false, updatable = false)
     @CreationTimestamp
     private LocalDateTime createdAt;
@@ -34,4 +31,5 @@ public class SteppingStonesScore {
     @Enumerated(EnumType.STRING)
     @Column(name = "status", nullable = false)
     private Status status = Status.ACTIVE;
+
 }
