+++ src/main/java/ku/user/domain/ranking/dto/response/SaveSteppingResponse.java
@@ -1,6 +1,6 @@
-package ku.user.domain.Ranking.dto.response;
+package ku.user.domain.ranking.dto.response;
 
-import ku.user.domain.Ranking.domain.SteppingStonesScore;
+import ku.user.domain.ranking.domain.SteppingStonesScore;
 
 public record SaveSteppingResponse(Long id, String nickName, int score, int coin) {
     public static SaveSteppingResponse fromEntity(SteppingStonesScore steppingStonesScore) {
