+++ src/main/java/ku/user/domain/ranking/service/SteppingStonesScoreService.java
@@ -1,8 +1,7 @@
-package ku.user.domain.Ranking.service;
+package ku.user.domain.ranking.service;
 
-import ku.user.domain.Ranking.domain.SteppingStonesScore;
+import ku.user.domain.ranking.domain.SteppingStonesScore;
 
-import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
